refactor!: make contractDeploy a wrapper, remove unused code
